Move bkwd_calculate_shifts method and helper to BkwdCalc module
